Extract reusable NewEpisodeBadge component with shimmer in core:designsystem and resolve code duplication in home and library features
